Refactor: searchBarComponent
