+++ src/components/searchPageComponent/menuBar/menuBarComponent.tsx
@@ -1,4 +1,4 @@
-import styles from '@/components/searchPage/menuBar/style';
+import styles from '@/components/searchPageComponent/menuBar/style';
 
 const MenuBarComponent = () => {
   return (
