+++ src/components/searchPageComponent/menuBar/style.ts
